+++ libkbfs/mocks_test.go
@@ -6754,75 +6754,6 @@ func (mr *MockRekeyQueueMockRecorder) Shutdown() *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Shutdown", reflect.TypeOf((*MockRekeyQueue)(nil).Shutdown))
 }
 
-// MockKeyBundleCache is a mock of KeyBundleCache interface
-type MockKeyBundleCache struct {
-	ctrl     *gomock.Controller
-	recorder *MockKeyBundleCacheMockRecorder
-}
-
-// MockKeyBundleCacheMockRecorder is the mock recorder for MockKeyBundleCache
-type MockKeyBundleCacheMockRecorder struct {
-	mock *MockKeyBundleCache
-}
-
-// NewMockKeyBundleCache creates a new mock instance
-func NewMockKeyBundleCache(ctrl *gomock.Controller) *MockKeyBundleCache {
-	mock := &MockKeyBundleCache{ctrl: ctrl}
-	mock.recorder = &MockKeyBundleCacheMockRecorder{mock}
-	return mock
-}
-
-// EXPECT returns an object that allows the caller to indicate expected use
-func (m *MockKeyBundleCache) EXPECT() *MockKeyBundleCacheMockRecorder {
-	return m.recorder
-}
-
-// GetTLFReaderKeyBundle mocks base method
-func (m *MockKeyBundleCache) GetTLFReaderKeyBundle(arg0 TLFReaderKeyBundleID) (*TLFReaderKeyBundleV3, error) {
-	ret := m.ctrl.Call(m, "GetTLFReaderKeyBundle", arg0)
-	ret0, _ := ret[0].(*TLFReaderKeyBundleV3)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// GetTLFReaderKeyBundle indicates an expected call of GetTLFReaderKeyBundle
-func (mr *MockKeyBundleCacheMockRecorder) GetTLFReaderKeyBundle(arg0 interface{}) *gomock.Call {
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetTLFReaderKeyBundle", reflect.TypeOf((*MockKeyBundleCache)(nil).GetTLFReaderKeyBundle), arg0)
-}
-
-// GetTLFWriterKeyBundle mocks base method
-func (m *MockKeyBundleCache) GetTLFWriterKeyBundle(arg0 TLFWriterKeyBundleID) (*TLFWriterKeyBundleV3, error) {
-	ret := m.ctrl.Call(m, "GetTLFWriterKeyBundle", arg0)
-	ret0, _ := ret[0].(*TLFWriterKeyBundleV3)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
-}
-
-// GetTLFWriterKeyBundle indicates an expected call of GetTLFWriterKeyBundle
-func (mr *MockKeyBundleCacheMockRecorder) GetTLFWriterKeyBundle(arg0 interface{}) *gomock.Call {
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetTLFWriterKeyBundle", reflect.TypeOf((*MockKeyBundleCache)(nil).GetTLFWriterKeyBundle), arg0)
-}
-
-// PutTLFReaderKeyBundle mocks base method
-func (m *MockKeyBundleCache) PutTLFReaderKeyBundle(arg0 TLFReaderKeyBundleID, arg1 TLFReaderKeyBundleV3) {
-	m.ctrl.Call(m, "PutTLFReaderKeyBundle", arg0, arg1)
-}
-
-// PutTLFReaderKeyBundle indicates an expected call of PutTLFReaderKeyBundle
-func (mr *MockKeyBundleCacheMockRecorder) PutTLFReaderKeyBundle(arg0, arg1 interface{}) *gomock.Call {
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "PutTLFReaderKeyBundle", reflect.TypeOf((*MockKeyBundleCache)(nil).PutTLFReaderKeyBundle), arg0, arg1)
-}
-
-// PutTLFWriterKeyBundle mocks base method
-func (m *MockKeyBundleCache) PutTLFWriterKeyBundle(arg0 TLFWriterKeyBundleID, arg1 TLFWriterKeyBundleV3) {
-	m.ctrl.Call(m, "PutTLFWriterKeyBundle", arg0, arg1)
-}
-
-// PutTLFWriterKeyBundle indicates an expected call of PutTLFWriterKeyBundle
-func (mr *MockKeyBundleCacheMockRecorder) PutTLFWriterKeyBundle(arg0, arg1 interface{}) *gomock.Call {
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "PutTLFWriterKeyBundle", reflect.TypeOf((*MockKeyBundleCache)(nil).PutTLFWriterKeyBundle), arg0, arg1)
-}
-
 // MockRekeyFSM is a mock of RekeyFSM interface
 type MockRekeyFSM struct {
 	ctrl     *gomock.Controller
